WIP - convert roadmap html elements to components
