+++ src/pages/roadmap/roadmaps-module.css
@@ -1,12 +1,12 @@
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Inter&display=swap');
 
-.square-title {
+.personal-roadmap-title {
     font-family: 'Poppins', sans-serif;
     font-weight: none;
 }
 
-.square {
+.personal-roadmap {
     margin: 10px;
     height: 20vh;
     display: inline-table;
@@ -14,20 +14,21 @@
     font-family: 'Inter', sans-serif;
 }
 
-.square:first-child {
-    margin: 10px 10px 10px 0;
-    height: 20vh;
-    display: inline-table;
-    background-color: #42B0FF;
-    font-family: 'Inter', sans-serif;
-}
-
-.squares-container {
+.personal-roadmap-container {
     overflow: auto;
-    white-space: nowrap;
     display: flex;
     overflow-y: hidden;
     margin-top: 1vh;
+    margin-left: 0;
+    background-color: #eeebeb;
+    padding: 1vw;
+    box-shadow: rgba(0, 0, 0, 0.15) 0px -2px 10px 0px;
+}
+
+.personal-roadmap-container-grid {
+    display: flex;
+    background-color: #eeebeb;
+    padding: 1vw;
 }
 
 .main-container {
@@ -55,50 +56,6 @@
     text-decoration: none;
     font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
     position: relative;
-    width: 20vw;
-    padding: 0.4vw;
-}
-
-.delete-button {
-    background-color: #ce1100;
-    color: #ffffff;
-    font-size: 22px;
-    font-weight: bold;
-    cursor: pointer;
-    text-decoration: none;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    position: relative;
-    border: none;
-}
-
-.roadmap-buttons {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-.edit-button {
-    background-color: #0048ce;
-    color: #ffffff;
-    font-size: 22px;
-    font-weight: bold;
-    cursor: pointer;
-    text-decoration: none;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    position: relative;
-    border: none;
-}
-
-.button-text {
-    color: #ffffff;
-    text-decoration: none;
-    vertical-align: middle;
-}
-
-.edit-roadmap-title {
-    font-size: 20px;
-    font-weight: bold;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    color: black;
-    text-align: center;
+    width: 15%;
+    padding: 0.1vw;
 }
